\name{ONTTermsAnnDbBimap}
\docType{class}

\alias{ONTTermsAnnDbBimap}
\alias{class:ONTTermsAnnDbBimap}
\alias{ONTTermsAnnDbBimap-class}



% Standard generic methods

\title{Class "ONTTermsAnnDbBimap" }

\description{
  A sub-class of  Bimap, specific for topOnto.xx.db. Please see \code{\link[AnnotationDbi:Bimap]{Bimap}} for details.
}


\seealso{
  \code{\link[AnnotationDbi:Bimap]{Bimap}},
  \code{\link{ONTTERM}}
}

\examples{
  class(ONTTERM)
}

\keyword{classes}
\keyword{interface}
